Move Sealevel-related transaction factor to proper module
