+++ rust/main/submitter/src/chain_tx_adapter/chains.rs
@@ -2,7 +2,6 @@ mod builder;
 mod cosmos;
 mod ethereum;
 mod sealevel;
-mod transaction;
 
 pub(crate) use builder::ChainTxAdapterFactory;
 pub(crate) use sealevel::SealevelPayload;
