+++ rust/main/submitter/src/chain_tx_adapter/chains/sealevel.rs
@@ -5,3 +5,4 @@ pub(crate) use precursor::SealevelTxPrecursor;
 mod adapter;
 mod payload;
 mod precursor;
+mod transaction;
