+++ rust/main/submitter/src/chain_tx_adapter/chains/sealevel/transaction.rs
@@ -7,7 +7,7 @@ use crate::transaction::{Transaction, TransactionId, TransactionStatus, VmSpecif
 pub struct TransactionFactory {}
 
 impl TransactionFactory {
-    pub fn new(payload: &FullPayload, precursor: SealevelTxPrecursor) -> Transaction {
+    pub fn build(payload: &FullPayload, precursor: SealevelTxPrecursor) -> Transaction {
         Transaction {
             id: TransactionId::new(Uuid::new_v4()),
             hash: None,
